modules/snowflake_aws_glue_s3_access/requirements.txt
cryptography>=45.0.5,<46.0.0
pyjwt>=2.10.1,<3.0.0